lib: update transport to draft 22

+++ include/quiche.h
@@ -236,7 +236,7 @@ uint64_t quiche_conn_timeout_as_nanos(quiche_conn *conn);
 void quiche_conn_on_timeout(quiche_conn *conn);
 
 // Closes the connection with the given error and reason.
-int quiche_conn_close(quiche_conn *conn, bool app, uint16_t err,
+int quiche_conn_close(quiche_conn *conn, bool app, uint64_t err,
                       const uint8_t *reason, size_t reason_len);
 
 // Returns the negotiated ALPN protocol.
